+++ pakyow-core/lib/pakyow/application/behavior/rescuing.rb
@@ -11,40 +11,52 @@ module Pakyow
       module Rescuing
         extend Support::Extension
 
-        # Error the app was rescued from.
-        #
-        attr_reader :rescued
+        common_methods do
+          # Error the app was rescued from.
+          #
+          def rescued
+            defined?(@rescued) ? @rescued : self.class.rescued
+          end
 
-        # Returns true if the app has been rescued.
-        #
-        def rescued?
-          instance_variable_defined?(:@rescued) && !!@rescued
+          # Returns true if the app has been rescued.
+          #
+          def rescued?
+            if is_a?(Application)
+              self.class.rescued? || (defined?(@rescued) && !!@rescued)
+            else
+              defined?(@rescued) && !!@rescued
+            end
+          end
+
+          # Enters rescue mode after logging the error.
+          #
+          private def rescue!(error)
+            @rescued = error
+
+            performing :rescue do
+              Pakyow.logger.error(error)
+
+              if is_a?(Application)
+                singleton_class.include Rescued
+              else
+                include Rescued
+              end
+            end
+          end
         end
 
-        private
-
-        # Enters rescue mode after logging the error.
-        #
-        def rescue!(error)
-          @rescued = error
-
-          performing :rescue do
-            Pakyow.logger.error(error)
-
+        module Rescued
+          def call(connection)
             message = <<~ERROR
               #{self.class} failed to initialize.
 
-              #{error.message}
-              #{error.backtrace.join("\n")}
+              #{rescued.message}
+              #{rescued.backtrace.join("\n")}
             ERROR
 
-            # Override call to always return an errored response.
-            #
-            define_singleton_method :call do |connection|
-              connection.status = 500
-              connection.body = StringIO.new(message)
-              connection.halt
-            end
+            connection.status = 500
+            connection.body = StringIO.new(message)
+            connection.halt
           end
         end
       end